+++ OrionUO/Network/Connection.h
@@ -9,7 +9,7 @@ public:
     bool GameSocket = false;
     bool UseProxy = false;
     string ProxyAddress = "";
-    int ProxyPort = 0;
+    uint16_t ProxyPort = 0;
     bool ProxySocks5 = false;
     string ProxyAccount = "";
     string ProxyPassword = "";
@@ -21,6 +21,6 @@ public:
     CSocket(bool gameSocket);
     ~CSocket();
 
-    virtual bool Connect(const string &address, int port);
+    virtual bool Connect(const string &address, uint16_t port);
     virtual vector<uint8_t> Decompression(vector<uint8_t> data);
 };
